requirements.txt
aiohttp>=3.3.0,<3.5.0
websockets>=6.0,<7.0
praw>=6.0.0
python-dotenv==0.9.1
-e git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]